add store.jsp and success store list

+++ src/main/java/com/sw/fd/repository/StoreRepository.java
@@ -2,5 +2,5 @@ package com.sw.fd.repository;
 import com.sw.fd.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Long> {
+public interface StoreRepository extends JpaRepository<Store, Integer> {
 }
